+++ controllers/timeslotsController.php
@@ -18,28 +18,25 @@ class TimeslotsController
 
     public function showTimeslots()
     {
+        if(Security::isLogged()){
         $data["timeslotList"] = $this->timeslot->getAll();
         View::render("timeslot/all", $data);
-        // if (Security::haySesion()) {
-        //     $data["timeslotList"] = $this->timeslot->getAll();
-        //     View::render("timeslot/all", $data);
-        // } else {
-        //     $data["error"] = "No tienes permiso para eso";
-        //     View::render("usuario/login", $data);
-        // }
+        } else {
+            $data["error"] = "No tienes permiso para eso";
+            View::render("user/login", $data);
+        }
     }
 
     // --------------------------------- FORMULARIO ALTA DE LIBROS ----------------------------------------
 
     public function formAddTimeslot()
     {   
+        if(Security::isLogged()){
         View::render("timeslot/form");
-        /* if (Security::haySesion()) {
-            View::render("timeslot/form", $data);
         } else {
             $data["error"] = "No tienes permiso para eso";
-            View::render("usuario/login", $data);
-        } */
+            View::render("user/login", $data);
+        }
     }
 
     // --------------------------------- INSERTAR LIBROS ----------------------------------------
@@ -56,57 +53,21 @@ class TimeslotsController
             $data["error"] = "No tienes permiso para eso";
             View::render("usuario/login", $data);
         }
-
-        /* if (Security::haySesion()) {
-            // Primero, recuperamos todos los datos del formulario
-            $titulo = Security::limpiar($_REQUEST["titulo"]);
-            $genero = Security::limpiar($_REQUEST["genero"]);
-            $pais = Security::limpiar($_REQUEST["pais"]);
-            $ano = Security::limpiar($_REQUEST["ano"]);
-            $numPaginas = Security::limpiar($_REQUEST["numPaginas"]);
-            $autores = Security::limpiar($_REQUEST["autor"]);
-
-            $result = $this->libro->insert($titulo, $genero, $pais, $ano, $numPaginas);
-            if ($result == 1) {
-                // Si la inserción del libro ha funcionado, continuamos insertando los autores, pero
-                // necesitamos conocer el id del libro que acabamos de insertar
-                $idLibro = $this->libro->getMaxId();
-                // Ya podemos insertar todos los autores junto con el libro en "escriben"
-                $result = $this->libro->insertAutores($idLibro, $autores);
-                if ($result > 0) {
-                    $data["info"] = "Libro insertado con éxito";
-                } else {
-                    $data["error"] = "Error al insertar los autores del libro";
-                }
-            } else {
-                // Si la inserción del libro ha fallado, mostramos mensaje de error
-                $data["error"] = "Error al insertar el libro";
-            }
-            $data["listaLibros"] = $this->libro->getAll();
-            View::render("libro/all", $data);
-        } else {
-            $data["error"] = "No tienes permiso para eso";
-            View::render("usuario/login", $data);
-        } */
     }
 
     // --------------------------------- BORRAR LIBROS ----------------------------------------
 
     public function deleteTimeslot()
     {   
-        $id = $_REQUEST["id"];
-        $result = $this->timeslot->delete($id);
-        $data["timeslotList"] = $this->timeslot->getAll();
-        header("Location: index.php?controller=TimeslotsController&action=showTimeslots");
-        /* if (Security::haySesion()) {
+        if(Security::isLogged()){
             $id = $_REQUEST["id"];
-            $result = $this->libro->delete($id);
-            $data["listaLibros"] = $this->libro->getAll();
-            View::render("libro/all", $data);
+            $result = $this->timeslot->delete($id);
+            $data["timeslotList"] = $this->timeslot->getAll();
+            header("Location: index.php?controller=TimeslotsController&action=showTimeslots");
         } else {
             $data["error"] = "No tienes permiso para eso";
-            View::render("usuario/login", $data);
-        } */
+            View::render("user/login", $data);
+        }
     }
